class hierarchy was changed
